--- project/.env.example
@@ -1,2 +1,0 @@
-PORT=9999
-DISABLE_ESLINT_PLUGIN=true
