+++ project/src/index.tsx
@@ -5,10 +5,11 @@ import App from './components/app/app';
 const Setting = {
   OFFERS_NUMBER: 6,
 };
+const offers = Array.from({ length:Setting.OFFERS_NUMBER }, (_, i:number) => i);
 
 ReactDOM.render(
   <React.StrictMode>
-    <App offersNumber={Setting.OFFERS_NUMBER} />
+    <App offers={offers} />
   </React.StrictMode>,
   document.getElementById('root'),
 );
