+++ project/src/pages/main-empty/main-empty.tsx
@@ -1,0 +1,29 @@
+import Header from '../../components/header/header';
+import Tabs from '../../components/tabs/tabs';
+
+function MainEmpty(): JSX.Element {
+  return (
+    <div className="page page--gray page--main">
+      <Header/>
+      <main className="page__main page__main--index page__main--index-empty">
+        <Tabs/>
+        <div className="cities">
+          <div className="cities__places-container cities__places-container--empty container">
+            <section className="cities__no-places">
+              <div className="cities__status-wrapper tabs__content">
+                <b className="cities__status">No places to stay available</b>
+                <p className="cities__status-description">
+                  We could not find any property available at the moment in
+                  Dusseldorf
+                </p>
+              </div>
+            </section>
+            <div className="cities__right-section"></div>
+          </div>
+        </div>
+      </main>
+    </div>
+  );
+}
+
+export default MainEmpty;
