+++ project/src/pages/property/property.tsx
@@ -1,37 +1,9 @@
-function PropertyNotLogged(): JSX.Element {
+import Header from '../../components/header/header';
+
+function Property(): JSX.Element {
   return (
     <div className="page">
-      <header className="header">
-        <div className="container">
-          <div className="header__wrapper">
-            <div className="header__left">
-              <a className="header__logo-link" href="main.html">
-                <img
-                  className="header__logo"
-                  src="img/logo.svg"
-                  alt="6 cities logo"
-                  width="81"
-                  height="41"
-                />
-              </a>
-            </div>
-            <nav className="header__nav">
-              <ul className="header__nav-list">
-                <li className="header__nav-item user">
-                  <a
-                    className="header__nav-link header__nav-link--profile"
-                    href="#"
-                  >
-                    <div className="header__avatar-wrapper user__avatar-wrapper"></div>
-                    <span className="header__login">Sign in</span>
-                  </a>
-                </li>
-              </ul>
-            </nav>
-          </div>
-        </div>
-      </header>
-
+      <Header/>
       <main className="page__main page__main--property">
         <section className="property">
           <div className="property__gallery-container container">
@@ -105,7 +77,7 @@ function PropertyNotLogged(): JSX.Element {
               </div>
               <div className="property__rating rating">
                 <div className="property__stars rating__stars">
-                  <span style={{ width: '80%' }} />
+                  <span style={{ width: '80%'}} />
                   <span className="visually-hidden">Rating</span>
                 </div>
                 <span className="property__rating-value rating__value">
@@ -191,7 +163,7 @@ function PropertyNotLogged(): JSX.Element {
                     <div className="reviews__info">
                       <div className="reviews__rating rating">
                         <div className="reviews__stars rating__stars">
-                          <span style={{ width: '80%' }} />
+                          <span style={{ width: '80%'}} />
                           <span className="visually-hidden">Rating</span>
                         </div>
                       </div>
@@ -206,6 +178,121 @@ function PropertyNotLogged(): JSX.Element {
                     </div>
                   </li>
                 </ul>
+                <form className="reviews__form form" action="#" method="post">
+                  <label
+                    className="reviews__label form__label"
+                    htmlFor="review"
+                  >
+                    Your review
+                  </label>
+                  <div className="reviews__rating-form form__rating">
+                    <input
+                      className="form__rating-input visually-hidden"
+                      name="rating"
+                      value="5"
+                      id="5-stars"
+                      type="radio"
+                    />
+                    <label
+                      htmlFor="5-stars"
+                      className="reviews__rating-label form__rating-label"
+                      title="perfect"
+                    >
+                      <svg className="form__star-image" width="37" height="33">
+                        <use xlinkHref="#icon-star" />
+                      </svg>
+                    </label>
+
+                    <input
+                      className="form__rating-input visually-hidden"
+                      name="rating"
+                      value="4"
+                      id="4-stars"
+                      type="radio"
+                    />
+                    <label
+                      htmlFor="4-stars"
+                      className="reviews__rating-label form__rating-label"
+                      title="good"
+                    >
+                      <svg className="form__star-image" width="37" height="33">
+                        <use xlinkHref="#icon-star" />
+                      </svg>
+                    </label>
+
+                    <input
+                      className="form__rating-input visually-hidden"
+                      name="rating"
+                      value="3"
+                      id="3-stars"
+                      type="radio"
+                    />
+                    <label
+                      htmlFor="3-stars"
+                      className="reviews__rating-label form__rating-label"
+                      title="not bad"
+                    >
+                      <svg className="form__star-image" width="37" height="33">
+                        <use xlinkHref="#icon-star" />
+                      </svg>
+                    </label>
+
+                    <input
+                      className="form__rating-input visually-hidden"
+                      name="rating"
+                      value="2"
+                      id="2-stars"
+                      type="radio"
+                    />
+                    <label
+                      htmlFor="2-stars"
+                      className="reviews__rating-label form__rating-label"
+                      title="badly"
+                    >
+                      <svg className="form__star-image" width="37" height="33">
+                        <use xlinkHref="#icon-star" />
+                      </svg>
+                    </label>
+
+                    <input
+                      className="form__rating-input visually-hidden"
+                      name="rating"
+                      value="1"
+                      id="1-star"
+                      type="radio"
+                    />
+                    <label
+                      htmlFor="1-star"
+                      className="reviews__rating-label form__rating-label"
+                      title="terribly"
+                    >
+                      <svg className="form__star-image" width="37" height="33">
+                        <use xlinkHref="#icon-star" />
+                      </svg>
+                    </label>
+                  </div>
+                  <textarea
+                    className="reviews__textarea form__textarea"
+                    id="review"
+                    name="review"
+                    placeholder="Tell how was your stay, what you like and what can be improved"
+                  />
+                  <div className="reviews__button-wrapper">
+                    <p className="reviews__help">
+                      To submit review please make sure to set{' '}
+                      <span className="reviews__star">rating</span> and describe
+                      your stay with at least{' '}
+                      <b className="reviews__text-amount">50 characters</b>.
+                    </p>
+                    <button
+                      className="reviews__submit form__submit button"
+                      type="submit"
+                      disabled={false}
+                    >
+                      Submit
+                    </button>
+                  </div>
+                </form>
               </section>
             </div>
           </div>
@@ -246,7 +333,7 @@ function PropertyNotLogged(): JSX.Element {
                         width="18"
                         height="19"
                       >
-                        <use xlinkHref="#icon-bookmark"></use>
+                        <use xlinkHref="#icon-bookmark" />
                       </svg>
                       <span className="visually-hidden">In bookmarks</span>
                     </button>
@@ -365,4 +452,4 @@ function PropertyNotLogged(): JSX.Element {
   );
 }
 
-export default PropertyNotLogged;
+export default Property;
